pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "bookmarks-cli"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.47.0",
    "python-frontmatter>=1.1.0",
    "trafilatura>=1.7.0",
    "httpx>=0.26.0",
]

[project.scripts]
bookmarks = "bookmarks.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["bookmarks*"]
